map reduce init

+++ src/main/java/com/banditos/dummymapreduce/FileWorker.java
@@ -1,31 +1,47 @@
 package com.banditos.dummymapreduce;
 
-import java.io.BufferedInputStream;
 import java.io.BufferedReader;
 import java.io.BufferedWriter;
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.OutputStream;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.List;
-import java.util.concurrent.Executor;
+import java.util.Map;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
+import java.util.concurrent.Future;
+import java.util.stream.Collectors;
 
 public class FileWorker {
 
-    Executor executor = Executors.newWorkStealingPool();
+    private int nThreads;
+    private ExecutorService executor;
+    private boolean toSplit;
 
-    public List<Path> splitFile(Path readFile, int splitAmount) throws IOException {
-        long fileSize = Files.size(readFile);
-        long chunkSize = fileSize / splitAmount;
+    public FileWorker(int nThreads, boolean toSplit) {
+        this.nThreads = nThreads;
+        this.executor = Executors.newFixedThreadPool(nThreads);
+        this.toSplit = toSplit;
+    }
 
+
+    public List<Path> splitFile(Path readFile) throws IOException {
         String parentPathStr = readFile.getParent().toString();
         Path dirPath = Paths.get(parentPathStr, "splitter");
+        if (!toSplit) {
+            return Arrays.asList(new File(dirPath.toUri()).listFiles())
+                    .stream()
+                    .map(file -> Paths.get(file.getPath()))
+                    .collect(
+                    Collectors.toList());
+        }
+
 
         for (File f : new File(dirPath.toUri()).listFiles()) {
             f.delete();
@@ -39,8 +55,10 @@ public class FileWorker {
         int i = 1;
         long ptr = 0;
         List<Path> paths = new ArrayList<>();
+        long fileSize = Files.size(readFile);
+        long chunkSize = fileSize / nThreads;
 
-        while (i <= splitAmount) {
+        while (i <= nThreads) {
             Path writeFile = Files.createFile(Paths.get(
                     parentPathStr, "splitter", "split" + ptr + ".txt"));
             paths.add(writeFile);
@@ -52,5 +70,24 @@ public class FileWorker {
             i++;
         }
         return paths;
+    }
+
+    public int countWordEntries(List<Path> files, String word) {
+        List<Future<Map<String, Integer>>> futures = new ArrayList<>();
+        for (int i = 0; i < nThreads; i++) {
+            final Future<Map<String, Integer>> future = executor
+                    .submit(new MapCounter(files.get(i)));
+            futures.add(future);
+        }
+        int total = 0;
+        for (Future<Map<String, Integer>> f : futures) {
+            try {
+                total += f.get().getOrDefault(word, 0);
+            } catch (InterruptedException | ExecutionException e) {
+                e.printStackTrace();
+            }
+        }
+        executor.shutdown();
+        return total;
     }
 }
